Support cascade replication in get_stats_replication_delays
When monitoring a cluster with cascade replication the current query
fails with "get_stats_replication_delays". We can instead monitor the
diff using pg_last_wal_receive_lsn, this will emit metrics showing how
much replication delay the replica has compared to the upstream node
(but not relative to the primary instance of the cluster).